refactor(subagent): compare the children count without a narrowing cast
The cap compared `running_children as u32` against the limit. Widening the
limit to `usize` instead is lossless in the direction that matters — a count
can exceed `u32::MAX` in principle, a `u32` limit never exceeds `usize` — so
the comparison no longer depends on the tree cap happening to bound the
count. No behavior change.

+++ crates/loop/ironclaw_loop_host/src/subagent_spawn_port.rs
@@ -787,7 +787,7 @@ impl SubagentSpawnCapabilityPort {
             .into_iter()
             .filter(|record| !record.status.is_terminal())
             .count();
-        if running_children as u32 >= self.limits.max_concurrent_children {
+        if running_children >= self.limits.max_concurrent_children as usize {
             return Ok(spawn_rejected("concurrent_children_cap_exceeded"));
         }
